requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==4.2.2
amqp==5.3.1
asgiref==3.8.1
aspose-words==25.3.0
async-timeout==4.0.3
attrs==24.1.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blinker==1.7.0
blosc2==2.7.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cryptography==44.0.0
cssbeautifier==1.15.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_ag_grid==31.2.0
dataframe_image==0.2.4
defusedxml==0.7.1
Django==4.2
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-debug-toolbar==4.3.0
django-environ==0.12.0
django-oauth-toolkit==3.0.1
django-plotly-dash==2.2.2
django-postgres-extra==2.0.8
django-redis==5.4.0
django-rest==0.8.7
django-select2==8.1.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djlint==1.34.1
docx==0.2.4
docxcompose==1.4.0
docxtpl==0.19.1
dpd_components==0.1.0
EditorConfig==0.12.4
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html2image==2.0.4.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
kaleido==0.2.1
kiwisolver==1.4.5
kombu==5.4.2
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.2
mkdocs==1.5.3
more-itertools==10.4.0
msgpack==1.0.8
multidict==6.0.5
mysql-connector==2.2.9
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.8
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
pathspec==0.12.1
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.0
plotly==5.22.0
prettytable==3.10.0
prompt_toolkit==3.0.48
protobuf==5.27.3
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pyTelegramBotAPI==4.22.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==26.1.1
redis==5.0.3
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
scipy==1.14.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.4.4
statsmodels==0.14.2
streamlit==1.37.1
streamlit-aggrid==1.0.5
tables==3.9.2
tabulate==0.9.0
tenacity==8.2.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urbansim==3.2
urllib3==2.2.2
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
XlsxWriter==3.2.0
yarl==1.9.4
